+++ .husky/pre-commit
@@ -2,11 +2,15 @@
 bun run lint-staged
 
 changed_files="$(git diff --cached --name-only)"
+
 if echo "$changed_files" | grep -Eq '\.langium$'; then
-  bun run --cwd packages/docml generate
+  bun run --cwd packages/language generate
 fi
 if echo "$changed_files" | grep -Eq '(src|test)\/'; then
   bun test --coverage
+fi
+if echo "$changed_files" | grep -Eq '^packages\/language\/'; then
+  bun run --cwd packages/language build
 fi
 if echo "$changed_files" | grep -Eq '^packages\/docml\/'; then
   bun run --cwd packages/docml build
