+++ .vscode/launch.json
@@ -11,7 +11,10 @@
       "request": "launch",
       "args": ["--extensionDevelopmentPath=${workspaceFolder}/packages/docml"],
       "sourceMaps": true,
-      "outFiles": ["${workspaceFolder}/packages/docml/dist/**/*.js"]
+      "outFiles": [
+        "${workspaceFolder}/packages/docml/dist/**/*.js",
+        "${workspaceFolder}/packages/language/dist/**/*.js"
+      ]
     },
     {
       "name": "Attach to Language Server",
@@ -22,6 +25,7 @@
       "sourceMaps": true,
       "outFiles": [
         "${workspaceFolder}/packages/docml/dist/**/*.js",
+        "${workspaceFolder}/packages/language/dist/**/*.js",
         "${workspaceFolder}/node_modules/langium"
       ]
     }
