+++ package.json
@@ -5,6 +5,7 @@
   "license": "MIT",
   "private": true,
   "workspaces": [
+    "packages/language",
     "packages/docml",
     "packages/docs"
   ],
