--- packages/docml/.vscodeignore
@@ -1,4 +1,0 @@
-**/*.{js,map,ts,tsbuildinfo,svg,docml,langium,local,vsix}
-**/tsconfig.*
-.npmignore
-langium-config.json
