+++ packages/docml/package.json
@@ -1,30 +1,35 @@
 {
   "name": "docml",
-  "description": "Lightweight Document Markup Language",
+  "description": "VSCode extension and command-line for Docml",
   "version": "0.1.2",
   "type": "module",
+  "files": [
+    "assets",
+    "dist/**/*.{cjs,js,ts,map}",
+    "src/**/*.ts",
+    "docml.tmLanguage.json",
+    "language-configuration.json",
+    "LICENSE"
+  ],
   "scripts": {
-    "build": "tsc -b src && bun run build:extension && bun run build:language",
+    "build": "tsc -b src && bun run build:extension && bun run build:server",
     "build:extension": "bun build --minify --target=node --format=cjs --external=vscode --outfile=dist/extension/main.min.cjs src/extension/main.ts",
-    "build:language": "bun build --minify --target=node --format=cjs --ignore-dce-annotations --outfile=dist/language/main.min.cjs src/language/main.ts",
-    "generate": "langium generate --mode=production",
-    "prepublishOnly": "bun run generate && bun run build",
+    "build:server": "bun build --minify --target=node --format=cjs --ignore-dce-annotations --outfile=dist/server/main.min.cjs src/server/main.ts",
+    "prepublishOnly": "bun run build",
     "vscode:package": "vsce package --no-dependencies",
-    "vscode:publish": "bun run prepublishOnly && vsce publish --no-dependencies"
+    "vscode:publish": "vsce publish --no-dependencies"
   },
   "dependencies": {
+    "@docml/language": "workspace:*",
     "chalk": "~5.4.1",
     "commander": "~13.1.0",
-    "langium": "~3.4.0",
-    "vscode-languageclient": "~9.0.1",
-    "vscode-languageserver": "~9.0.1"
+    "vscode-languageclient": "~9.0.1"
   },
   "devDependencies": {
     "@tsconfig/bun": "^1.0.7",
     "@types/bun": "~1.2.5",
     "@types/vscode": "~1.67.0",
     "@vscode/vsce": "^3.2.2",
-    "langium-cli": "~3.4.0",
     "typescript": "~5.8.2"
   },
   "volta": {
@@ -87,26 +92,16 @@
   },
   "homepage": "https://dsogari.github.io/docml/",
   "keywords": [
-    "lightweight",
-    "document",
-    "markup",
-    "language"
+    "extension",
+    "vscode"
   ],
   "license": "MIT",
   "bin": {
     "docml": "dist/command/main.js"
   },
   "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    },
-    "./browser": {
-      "types": "./dist/browser/main.d.ts",
-      "default": "./dist/browser/main.js"
-    },
+    ".": null,
     "./tmLanguage": "./docml.tmLanguage.json",
     "./langConfig": "./language-configuration.json"
-  },
-  "types": "./dist/index.d.ts"
+  }
 }
