+++ packages/docml/src/extension/main.ts
@@ -26,7 +26,7 @@ export function deactivate(): Promise<void> | undefined {
 
 /** @ignore */
 function startLanguageClient(context: vscode.ExtensionContext): LanguageClient {
-  const serverModule = context.asAbsolutePath(path.join('dist', 'language', 'main.min.cjs'));
+  const serverModule = context.asAbsolutePath(path.join('dist', 'server', 'main.min.cjs'));
   // The debug options for the server
   // --inspect=6009: runs the server in Node's Inspector mode so VS Code can attach to the server for debugging.
   // By setting `process.env.DEBUG_BREAK` to a truthy value, the language server will wait until a debugger is attached.
