--- packages/docml/src/index.ts
@@ -1,2 +1,0 @@
-export { createDocmlServices } from './language/docml-module.js';
-export * from './language/generated/ast.js';
