+++ packages/docml/src/server/main.ts
@@ -1,0 +1,1 @@
+import '@docml/language/server';
