+++ packages/docs/package.json
@@ -14,6 +14,7 @@
     "dev:debug": "vite --debug --force"
   },
   "devDependencies": {
+    "@docml/language": "workspace:*",
     "docml": "workspace:*",
     "local-web-server": "^5.4.0",
     "monaco-editor-wrapper": "~6.5.0",
