+++ packages/language/assets/logo-wide.png
