+++ packages/language/src/index.ts
@@ -1,0 +1,3 @@
+export * from './language/docml-module.js';
+export * from './language/generated/ast.js';
+export * from './language/generated/module.js';
