+++ packages/language/tsconfig.base.json
@@ -1,0 +1,20 @@
+{
+  "extends": "@tsconfig/bun/tsconfig.json",
+  "compilerOptions": {
+    "types": ["bun"],
+    "lib": ["ESNext", "DOM", "WebWorker"],
+    "allowJs": false, // https://stackoverflow.com/a/42691024/4725347
+    "noEmit": false,
+    "sourceMap": true,
+    "incremental": true,
+    "stripInternal": true,
+    "declarationMap": true,
+    "esModuleInterop": true,
+    "noImplicitReturns": true,
+    "noImplicitOverride": true,
+    "verbatimModuleSyntax": false, // https://github.com/microsoft/TypeScript/issues/40344
+    "allowImportingTsExtensions": false,
+    "forceConsistentCasingInFileNames": true,
+    "disableSourceOfProjectReferenceRedirect": true
+  }
+}
